+++ packages/db/src/lib/granule.ts
@@ -11,6 +11,7 @@ import Logger from '@cumulus/logger';
 
 import { CollectionPgModel } from '../models/collection';
 import { GranulePgModel } from '../models/granule';
+import { createRejectableTransaction } from '../database';
 import { GranulesExecutionsPgModel } from '../models/granules-executions';
 import { PostgresGranule, PostgresGranuleRecord } from '../types/granule';
 import { GranuleWithProviderAndCollectionInfo } from '../types/query';
@@ -18,6 +19,8 @@ import { UpdatedAtRange } from '../types/record';
 const { deprecate } = require('@cumulus/common/util');
 
 const { TableNames } = require('../tables');
+
+const log = new Logger({ sender: '@cumulus/db/lib/granules' });
 
 export const getGranuleCollectionId = async (
   knexOrTransaction: Knex,
@@ -353,4 +356,29 @@ export const getGranulesByGranuleId = async (
   const records: PostgresGranuleRecord[] = await knexOrTransaction(granulesTable)
     .where({ granule_id: granuleId });
   return records;
+};
+
+/**
+ * Get granules from table where granule_id matches provided granuleId
+ *
+ * @param {Knex} knex - DB client or transaction
+ * @param {string} granuleId - Granule ID
+ * @returns {Promise<PostgresGranuleRecord[]>} The returned list of records
+ */
+export const updateBatchGranulesCollection = async (
+  knex: Knex,
+  granuleIds: Array<String>,
+  collectionCumulusId: number
+): Promise<void> => {
+  const {
+    granules: granulesTable,
+  } = TableNames;
+  try {
+    await createRejectableTransaction(knex, async (trx) => {
+      await trx(granulesTable).whereIn('granule_id', granuleIds).update({ collection_cumulus_id: collectionCumulusId });
+    });
+  } catch (thrownError) {
+    log.error(`Write Granules failed: ${JSON.stringify(thrownError)}`);
+    throw thrownError;
+  }
 };